requirements.txt
fastapi
uvicorn
google-adk
google-cloud-storage
google-api-python-client
google-auth
google-auth-oauthlib